hostapd: Change drv_flags from unsigned int to u64
Some flag already using a bit larger than 32, so extend the hostapd
drv_flags type similarly to the earlier wpa_supplicant change to get the
full flag content.


Git-commit: e3a8ad44055d5b453c7f080958137eb718a60b6e
Git-repo : git://w1.fi/srv/git/hostap.git
CRs-fixed: 752061

+++ src/ap/hostapd.h
@@ -288,7 +288,7 @@ struct hostapd_iface {
 	struct ap_info *ap_list; /* AP info list head */
 	struct ap_info *ap_hash[STA_HASH_SIZE];
 
-	unsigned int drv_flags;
+	u64 drv_flags;
 
 	/*
 	 * A bitmap of supported protocols for probe response offload. See
